refactor(frontend): update editUser dialog
